// src/types/types.ts
interface ProductImage {
  imageUrl: string;
  public_id: string;
  altText?: string;
  imageIndex?: number;

  size?: string;

  color?: string;
}
export interface IProduct {
  id: number;
  name: string;
  price: number;
  description: string;
  stock: number;
  discountPrice: number;
  sale?: string;
  colors?: [];
  spacification?: specsDetails[];
  posterImageUrl: string | StaticImageData;
  posterImagePublicId: string;
  posterImageAltText?: string;
  hoverImageUrl: string;
  hoverImagePublicId: string;
  productImages: ProductImage[];
  additionalInformation: AdditionalInformation[];
  categoriesId: number;
  categories?: ICategory[];
  subcategories?: ICategory[];
  sections?: [];
  createdAt?: string;
  sale_counter?: string;
  sortedSubcategories?: ICategory[];
  sizes?: string[];
  filter?: Filter[];
}
interface Filter {
  heading: string;
  additionalInformation: {
    name: string;
    price: string;
  }[];
}
export interface specsDetails {
  id: number;
  specsDetails: string;
}
// ...
export interface IProductDetail {
  name: string;
}

export interface ITabbyList {
  id: number;
  para: string;
}
export interface ITabbyPayList {
  id: number;
  imageUrl: StaticImageData;
}

export interface ITamaraList {
  id: number;
  title?: string;
  para: string;
}

export interface IMAGE_INTERFACE {
  public_id?: string;
  imageUrl?: string;
  name?: string;
}

export interface ICategory {
  id: number;
  name: string;
  createdAt?: string;
  posterImageUrl?: string;
  posterImagePublicId?: string;
  categories?: any;
  description?: string;
  short_description?: string;
  Images_Alt_Text?: string;
  Meta_Title?: string;
  Meta_Description?: string;
  Canonical_Tag?: string;
  subcategories?: ICategory[];
}
// ...
export interface IReview {
  id: number;
  name: string;
  email: string;
  review: string;
  star: number;
  createdAt: string;
  productId: number;
  userProfileImg?: string;
}

export type IProductWithoutId = Omit<IProduct, 'id'>;

export interface AdditionalInformation {
  key?: string;
  value?: string;
  colors?: string[];
  dimension?: string[];
}